actualizado con clase 10 28/05

+++ app/src/main/java/com/example/weatherapp/presentacion/clima/ClimaEstado.kt
@@ -3,9 +3,9 @@ package com.example.weatherapp.presentacion.clima
 sealed class ClimaEstado{
     data class Exitoso(
         val ciudad: String = "",
-        val temperatura: Int = 0,
+        val temperatura: Double = 0.0,
         val descripcion: String = "",
-        val st: Int = 0,
+        val st: Double = 0.0,
         val noHayDatos: Boolean = true
     ) : ClimaEstado()
     data class Error(
